[update] 찜목록 조회request -> title, cover, price 매핑 후 response

+++ Back/Dalcheon/Dalcheon/src/main/java/Bukgu/Dalcheon/domain/OpenApi/ItemDTO.java
@@ -16,8 +16,8 @@ public class ItemDTO {
     private String description;
     private String isbn13;
     private long itemId;
-    private int priceSales;
-    private int priceStandard;
+    private long priceSales;
+    private long priceStandard;
     private String cover;
     private int categoryId;
     private String categoryName;
